+++ wondrous-app/components/Common/UserAboutInfo/styles.tsx
@@ -1,10 +1,10 @@
 import { Box, Button, Dialog, IconButton, Typography } from '@mui/material';
 import styled from 'styled-components';
-import { LogoSquare } from '../ci';
-import { ModalCloseButton } from '../ModalCloseButton';
-import { Task } from '../Task';
-import { TaskInner, TaskWrapper } from '../Task/styles';
-import PodIcon from '../../Icons/podIcon';
+import { LogoSquare } from 'components/Common/ci';
+import { ModalCloseButton } from 'components/Common/ModalCloseButton';
+import Task from 'components/Common/Task';
+import { TaskInner, TaskWrapper } from 'components/Common/Task/styles';
+import PodIcon from 'components/Icons/podIcon';
 
 export const UserAboutInfoContainer = styled.div`
   width: 100%;
