+++ wondrous-app/components/ProfileUserTaskDaos/ProfileUserTaskDaos.tsx
@@ -1,4 +1,4 @@
-import { Task } from 'components/Common/Task';
+import Task from 'components/Common/Task';
 import OrgCard from 'components/OrgCard';
 import ProfileContentGrid from 'components/ProfileContentGrid';
 import useGetUserAboutPage from 'hooks/useGetUserAboutPage';
